Move global variables to `Functions.kt`
Anyone thinking global vars should not be used? I don't care haha... 
unless ... feel free to fork this repository and provide a better way of 
doing this ^^
